backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.1
numpy==1.26.4
scipy==1.13.0
python-multipart==0.0.9
pydantic==2.6.4
yfinance==0.2.37
sentence-transformers==2.7.0
faiss-cpu==1.8.0
ollama==0.2.1